+++ storybook/storybook/src/packages/Form/steps/src/Steps.stories.js
@@ -5,6 +5,8 @@ import { text, select } from '@storybook/addon-knobs';
 import addToStorie from '@axa-fr/storybook-addons';
 import { Steps, Step, StepBase, StepModes } from '@axa-fr/react-toolkit-form-steps';
 
+const readme = require('@axa-fr/react-toolkit-form-steps/dist/README.md');
+
 const stories = [];
 
 stories.push({
@@ -102,6 +104,7 @@ stories.push({
 
 const storyData = {
   name: 'Form.Steps',
+  docs:readme,
   stories,
 };