+++ storybook/storybook/src/packages/Layout/footer-client/src/FooterClientSimple.stories.js
@@ -8,6 +8,8 @@ import {
   FooterClientList,
 } from '@axa-fr/react-toolkit-layout-footer-client';
 
+const readme = require('@axa-fr/react-toolkit-layout-footer-client/dist/README.md');
+
 const stories = [];
 const columns = [
   {
@@ -92,6 +94,7 @@ stories.push({
 
 const storyData = {
   name: 'Layout.FooterClient',
+  docs: readme,
   stories,
 };