+++ storybook/storybook/src/packages/Layout/footer/src/FooterCore.stories.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import { action } from '@storybook/addon-actions';
 
 import addToStorie from '@axa-fr/storybook-addons';
 import { FooterCore } from '@axa-fr/react-toolkit-layout-footer';
@@ -8,10 +7,10 @@ const readme = require('@axa-fr/react-toolkit-layout-footer/dist/README.md');
 
 const storyData = {
   name: 'Layout.Footer',
+  docs: readme,
   stories: [
     {
       desc: 'Core with HTML children',
-      docs: readme,
       component: () => (
         <FooterCore icon="images/logo-axa.svg">
           <a href="https://www.axa.fr/">