+++ storybook/storybook/src/Home/CHANGELOG.md
@@ -2,12 +2,14 @@
 All notable changes to this project will be documented in this file.
 
 #### 1.0.0
-> 22 November 2018
+> 29 November 2018
+- doc(readme) add readme inside packages for npm [`#10`](https://github.com/AxaGuilDEv/react-toolkit/pull/10) by 
+- fix(demo) missing package [`#7`](https://github.com/AxaGuilDEv/react-toolkit/pull/7) by 
+- feat(demo) add example demo App [`#6`](https://github.com/AxaGuilDEv/react-toolkit/pull/6) by 
 - doc(readme) update links to website documentation [`#5`](https://github.com/AxaGuilDEv/react-toolkit/pull/5) by 
 - fix: update package-lock and delete page contrib on style demo [`#4`](https://github.com/AxaGuilDEv/react-toolkit/pull/4) by 
 - doc: fix COC link from README [`#3`](https://github.com/AxaGuilDEv/react-toolkit/pull/3) by 
 - doc(readme) add quality badge [`#2`](https://github.com/AxaGuilDEv/react-toolkit/pull/2) by 
-- Example demo App [`1585e42`](https://github.com/AxaGuilDEv/react-toolkit/commit/1585e42819ecf6613c5925f25f4b6ec08eca9e1e) by Samuel Gomez 
 - doc(readme-npm-badge) add npm badge [`c464b44`](https://github.com/AxaGuilDEv/react-toolkit/commit/c464b44661c9eabac18cadcd0d482e6ebef6f195) by Guillaume Chervet 
 - chore(release): publish v1.0.0 [ci skip] [`b007a14`](https://github.com/AxaGuilDEv/react-toolkit/commit/b007a14ebdc8395fee8cfde893ea432c4ada9c67) by Travis CI User 
 - chore(release): publish v0.0.2-alpha.0 [ci skip] [`00d6d80`](https://github.com/AxaGuilDEv/react-toolkit/commit/00d6d80d9609c4c4855f00a9dd89dd2b9125eb06) by Travis CI User